Import cleanup and rename to 'quote calculator'
